CompleteIntake action
