+++ client/app/intake/actions/ama.js
@@ -1,0 +1,51 @@
+import { ACTIONS, ENDPOINT_NAMES } from '../constants';
+import ApiUtil from '../../util/ApiUtil';
+import { formatRatingData } from '../util';
+
+const analytics = true;
+
+export const completeIntake = (intakeId, intakeData) => (dispatch) => {
+  dispatch({
+    type: ACTIONS.COMPLETE_INTAKE_START,
+    meta: { analytics }
+  });
+
+  const data = formatRatingData(intakeData);
+
+  return ApiUtil.patch(`/intake/${intakeId}/complete`, { data }, ENDPOINT_NAMES.COMPLETE_INTAKE).
+    then(
+      (response) => {
+        const responseObject = JSON.parse(response.text);
+
+        dispatch({
+          type: ACTIONS.COMPLETE_INTAKE_SUCCEED,
+          payload: {
+            intake: responseObject
+          },
+          meta: { analytics }
+        });
+
+        return true;
+      },
+      (error) => {
+        let responseObject = {};
+
+        try {
+          responseObject = JSON.parse(error.response.text);
+        } catch (ex) { /* pass */ }
+
+        const responseErrorCode = responseObject.error_code;
+        const responseErrorData = responseObject.error_data;
+
+        dispatch({
+          type: ACTIONS.COMPLETE_INTAKE_FAIL,
+          payload: {
+            responseErrorCode,
+            responseErrorData
+          },
+          meta: { analytics }
+        });
+        throw error;
+      }
+    );
+};
