+++ client/app/intake/actions/appeal.js
@@ -1,6 +1,5 @@
 import { ACTIONS, ENDPOINT_NAMES } from '../constants';
 import ApiUtil from '../../util/ApiUtil';
-import { formatRatingData } from '../util';
 import { formatDateStringForApi } from '../../util/DateUtil';
 import _ from 'lodash';
 
@@ -57,52 +56,6 @@ export const submitReview = (intakeId, appeal) => (dispatch) => {
           }
         });
 
-        throw error;
-      }
-    );
-};
-
-export const completeIntake = (intakeId, appeal) => (dispatch) => {
-  dispatch({
-    type: ACTIONS.COMPLETE_INTAKE_START,
-    meta: { analytics }
-  });
-
-  const data = formatRatingData(appeal);
-
-  return ApiUtil.patch(`/intake/${intakeId}/complete`, { data }, ENDPOINT_NAMES.COMPLETE_INTAKE).
-    then(
-      (response) => {
-        const responseObject = JSON.parse(response.text);
-
-        dispatch({
-          type: ACTIONS.COMPLETE_INTAKE_SUCCEED,
-          payload: {
-            intake: responseObject
-          },
-          meta: { analytics }
-        });
-
-        return true;
-      },
-      (error) => {
-        let responseObject = {};
-
-        try {
-          responseObject = JSON.parse(error.response.text);
-        } catch (ex) { /* pass */ }
-
-        const responseErrorCode = responseObject.error_code;
-        const responseErrorData = responseObject.error_data;
-
-        dispatch({
-          type: ACTIONS.COMPLETE_INTAKE_FAIL,
-          payload: {
-            responseErrorCode,
-            responseErrorData
-          },
-          meta: { analytics }
-        });
         throw error;
       }
     );
