+++ client/app/intake/actions/higherLevelReview.js
@@ -1,7 +1,6 @@
 import { ACTIONS, ENDPOINT_NAMES } from '../constants';
 import ApiUtil from '../../util/ApiUtil';
 import { formatDateStringForApi } from '../../util/DateUtil';
-import { formatRatingData } from '../util';
 import _ from 'lodash';
 
 const analytics = true;
@@ -70,52 +69,6 @@ export const submitReview = (intakeId, higherLevelReview) => (dispatch) => {
           }
         });
 
-        throw error;
-      }
-    );
-};
-
-export const completeIntake = (intakeId, higherLevelReview) => (dispatch) => {
-  dispatch({
-    type: ACTIONS.COMPLETE_INTAKE_START,
-    meta: { analytics }
-  });
-
-  const data = formatRatingData(higherLevelReview);
-
-  return ApiUtil.patch(`/intake/${intakeId}/complete`, { data }, ENDPOINT_NAMES.COMPLETE_INTAKE).
-    then(
-      (response) => {
-        const responseObject = JSON.parse(response.text);
-
-        dispatch({
-          type: ACTIONS.COMPLETE_INTAKE_SUCCEED,
-          payload: {
-            intake: responseObject
-          },
-          meta: { analytics }
-        });
-
-        return true;
-      },
-      (error) => {
-        let responseObject = {};
-
-        try {
-          responseObject = JSON.parse(error.response.text);
-        } catch (ex) { /* pass */ }
-
-        const responseErrorCode = responseObject.error_code;
-        const responseErrorData = responseObject.error_data;
-
-        dispatch({
-          type: ACTIONS.COMPLETE_INTAKE_FAIL,
-          payload: {
-            responseErrorCode,
-            responseErrorData
-          },
-          meta: { analytics }
-        });
         throw error;
       }
     );
