+++ client/app/intake/pages/higherLevelReview/finish.jsx
@@ -6,7 +6,7 @@ import RatedIssueCounter from '../../components/RatedIssueCounter';
 import NonRatedIssues from './nonRatedIssues';
 import RatedIssues from './ratedIssues';
 import { Redirect } from 'react-router-dom';
-import { completeIntake } from '../../actions/higherLevelReview';
+import { completeIntake } from '../../actions/ama';
 import { bindActionCreators } from 'redux';
 import { REQUEST_STATE, PAGE_PATHS, INTAKE_STATES } from '../../constants';
 import { getIntakeStatus } from '../../selectors';
